requirements.txt

nvidia-cublas-cu12==12.6.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.6.80
    # via torch
nvidia-cuda-nvrtc-cu12==12.6.77
    # via torch
nvidia-cuda-runtime-cu12==12.6.77
    # via torch
nvidia-cudnn-cu12==9.5.1.17
    # via torch
nvidia-cufft-cu12==11.3.0.4
    # via torch
nvidia-cufile-cu12==1.11.1.6
    # via torch
nvidia-curand-cu12==10.3.7.77
    # via torch
nvidia-cusolver-cu12==11.7.1.2
    # via torch
nvidia-cusparse-cu12==12.5.4.2
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.3
    # via torch
nvidia-nccl-cu12==2.26.2
    # via torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.6.77
    # via torch
torch==2.7.1
    # via
    #   persona-forge (pyproject.toml)
    #   sentence-transformers

aiofiles
aiosqlite
altair
annotated-types
antlr4-python3-runtime
anyio
asttokens
attrs
authlib
cffi
click
ctranslate2
deepagents>=0.0.3
dnspython
docstring-parser
docutils
email-validator
fastapi
fastmcp
ffmpy
filelock
fonttools
fsspec
gradio == 5.47.1
gradio-client == 1.13.2
groovy
h11
hf-xet
httpcore
httpx>=0.28.1
httpx-sse
huggingface-hub
idna
ipython
ipython-pygments-lexers
isodate
jedi
jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-path
jsonschema-specifications
kiwisolver
langchain-core
langchain-mcp-adapters
langchain-ollama
langgraph-checkpoint
langgraph-checkpoint-sqlite
langgraph-prebuilt
langgraph>=0.6.4
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langchain_community>=0.3.27
langchain_tavily>=0.2.7
langchain_mcp_adapters>=0.1.9
langchain_google_genai
langgraph-sdk
langsmith
lazy-object-proxy
markdownify>=1.2.0
markdown-it-py
markupsafe
matplotlib
matplotlib-inline
mcp
mdurl
more-itertools
mpmath
mypy>=1.11.1
narwhals
neo4j
networkx
numpy
ollama
openai
openapi-core
openapi-pydantic
openapi-schema-validator
openapi-spec-validator
orjson
ormsgpack
packaging
pandas
pillow
platformdirs
plotly
prompt-toolkit
ptyprocess
pure-eval
pycparser
pydantic>=2.0.0
pydantic-core
pydantic-settings
pydub
pygments
pyparsing
pyperclip
python-dateutil
python-dotenv
python-multipart
pytz
pyyaml
referencing
regex
requests
requests-toolbelt
rich>=2.0.0
ruff>=0.6.1
safehttpx
safetensors
scikit-learn
scipy
seaborn
semantic-version
sentence-transformers
setuptools
sqlite-vec
sympy
tavily-python>=0.5.0
tenacity
threadpoolctl
tokenizers
tomlkit
torch
typer
types-pytz
typing-extensions
typing-inspection
tzdata
urllib3
uvicorn
faster-whisper
